+++ pallets/relayer/src/tests.rs
@@ -6,19 +6,17 @@ use frame_support::{
     traits::Currency,
     BoundedVec,
 };
-use pallet_programs::AllowedToModifyProgram;
+use pallet_programs::ProgramInfo;
 use pallet_relayer::Call as RelayerCall;
 use sp_runtime::{
-    traits::SignedExtension,
+    traits::{Hash, SignedExtension},
     transaction_validity::{TransactionValidity, ValidTransaction},
 };
 
 use crate as pallet_relayer;
-use crate::{mock::*, Error, RegisteredInfo, RegisteringDetails, ValidateConfirmRegistered};
-
-/// consts used for testing
-const PROGRAM_MODIFICATION_ACCOUNT: u64 = 1u64;
-const SIG_REQ_ACCOUNT: u64 = 2u64;
+use crate::{
+    mock::*, Error, Registered, RegisteredInfo, RegisteringDetails, ValidateConfirmRegistered,
+};
 
 #[test]
 fn it_tests_get_validator_rotation() {
@@ -58,41 +56,19 @@ fn it_tests_get_validator_rotation() {
 fn it_registers_a_user() {
     new_test_ext().execute_with(|| {
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
 
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(1),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Public,
-            empty_program,
+            program_hash,
         ));
         assert_eq!(Relayer::dkg(0), vec![1u64.encode()]);
-    });
-}
-
-#[test]
-fn it_takes_a_program_storage_deposit_during_register() {
-    new_test_ext().execute_with(|| {
-        use frame_support::traits::Currency;
-
-        let program = vec![1u8, 2u8];
-        let initial_balance = 100;
-
-        Balances::make_free_balance_be(&PROGRAM_MODIFICATION_ACCOUNT, initial_balance);
-
-        assert_ok!(Relayer::register(
-            RuntimeOrigin::signed(SIG_REQ_ACCOUNT),
-            PROGRAM_MODIFICATION_ACCOUNT,
-            KeyVisibility::Public,
-            program.clone(),
-        ));
-
-        let expected_reserve = <Test as pallet_programs::Config>::ProgramDepositPerByte::get()
-            * (program.len() as u32);
-
-        assert_eq!(
-            Balances::free_balance(PROGRAM_MODIFICATION_ACCOUNT),
-            initial_balance - (expected_reserve as u64)
-        );
     });
 }
 
@@ -113,11 +89,17 @@ fn it_confirms_registers_a_user() {
         );
 
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(1),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Private([0; 32]),
-            empty_program,
+            program_hash.clone(),
         ));
 
         assert_noop!(
@@ -152,11 +134,11 @@ fn it_confirms_registers_a_user() {
         );
 
         let registering_info = RegisteringDetails::<Test> {
-            program_modification_account: 2 as <Test as frame_system::Config>::AccountId,
             confirmations: vec![0],
-            program: vec![],
+            program_pointer: program_hash.clone(),
             key_visibility: KeyVisibility::Private([0; 32]),
             verifying_key: Some(expected_verifying_key.clone()),
+            program_modification_account: 2,
         };
 
         assert_eq!(Relayer::registering(1), Some(registering_info));
@@ -173,12 +155,41 @@ fn it_confirms_registers_a_user() {
             Relayer::registered(1).unwrap(),
             RegisteredInfo {
                 key_visibility: KeyVisibility::Private([0; 32]),
-                verifying_key: expected_verifying_key
+                verifying_key: expected_verifying_key,
+                program_pointer: program_hash,
+                program_modification_account: 2
             }
         );
+    });
+}
 
-        // make sure program and sig req keys are set
-        assert!(AllowedToModifyProgram::<Test>::contains_key(2, 1));
+#[test]
+fn it_changes_a_program_pointer() {
+    new_test_ext().execute_with(|| {
+        let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
+        let new_program = vec![10];
+        let new_program_hash = <Test as frame_system::Config>::Hashing::hash(&new_program);
+        let expected_verifying_key = BoundedVec::default();
+
+        let mut registered_info = RegisteredInfo {
+            key_visibility: KeyVisibility::Public,
+            verifying_key: expected_verifying_key,
+            program_pointer: program_hash,
+            program_modification_account: 2,
+        };
+
+        Registered::<Test>::insert(1, registered_info.clone());
+        assert_eq!(Relayer::registered(1).unwrap(), registered_info.clone());
+
+        assert_ok!(Relayer::change_program_pointer(RuntimeOrigin::signed(2), 1, new_program_hash,));
+        registered_info.program_pointer = new_program_hash;
+        assert_eq!(Relayer::registered(1).unwrap(), registered_info);
     });
 }
 
@@ -186,13 +197,19 @@ fn it_confirms_registers_a_user() {
 fn it_fails_on_non_matching_verifying_keys() {
     new_test_ext().execute_with(|| {
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         let expected_verifying_key = BoundedVec::default();
         let unexpected_verifying_key = vec![10];
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(1),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Private([0; 32]),
-            empty_program,
+            program_hash,
         ));
         pallet_staking_extension::ThresholdToStash::<Test>::insert(1, 1);
         pallet_staking_extension::ThresholdToStash::<Test>::insert(2, 2);
@@ -222,17 +239,47 @@ fn it_doesnt_allow_double_registering() {
     new_test_ext().execute_with(|| {
         // register a user
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(1),
             2,
             KeyVisibility::Permissioned,
-            empty_program,
+            program_hash.clone(),
         ));
 
         // error if they try to submit another request, even with a different program key
         assert_noop!(
-            Relayer::register(RuntimeOrigin::signed(1), 2, KeyVisibility::Permissioned, vec![]),
+            Relayer::register(
+                RuntimeOrigin::signed(1),
+                2,
+                KeyVisibility::Permissioned,
+                program_hash
+            ),
             Error::<Test>::AlreadySubmitted
+        );
+    });
+}
+
+#[test]
+fn it_fails_no_program() {
+    new_test_ext().execute_with(|| {
+        // register a user
+        let non_existing_program = vec![10];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&non_existing_program);
+
+        assert_noop!(
+            Relayer::register(
+                RuntimeOrigin::signed(1),
+                2,
+                KeyVisibility::Permissioned,
+                program_hash
+            ),
+            Error::<Test>::ProgramDoesNotExist
         );
     });
 }
@@ -241,31 +288,41 @@ fn it_doesnt_allow_double_registering() {
 fn it_tests_prune_registration() {
     new_test_ext().execute_with(|| {
         let inital_program = vec![10];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&inital_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: inital_program, program_modification_account: 1 },
+        );
+
         Balances::make_free_balance_be(&2, 100);
         // register a user
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(1),
             2,
             KeyVisibility::Permissioned,
-            inital_program,
+            program_hash,
         ));
-        assert_eq!(Balances::free_balance(2), 95, "Deposit is charged");
         assert!(Relayer::registering(1).is_some(), "Make sure there is registering state");
         assert_ok!(Relayer::prune_registration(RuntimeOrigin::signed(1)));
         assert_eq!(Relayer::registering(1), None, "Make sure registering is pruned");
-        assert_eq!(Balances::free_balance(2), 100, "Deposit is returned");
     });
 }
 #[test]
 fn it_provides_free_txs_confirm_done() {
     new_test_ext().execute_with(|| {
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         let expected_verifying_key = BoundedVec::default();
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(5),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Public,
-            empty_program,
+            program_hash,
         ));
         let p = ValidateConfirmRegistered::<Test>::new();
         let c = RuntimeCall::Relayer(RelayerCall::confirm_register {
@@ -322,12 +379,18 @@ fn it_provides_free_txs_confirm_done_fails_2() {
 fn it_provides_free_txs_confirm_done_fails_3() {
     new_test_ext().execute_with(|| {
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         let expected_verifying_key = BoundedVec::default();
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(5),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Public,
-            empty_program,
+            program_hash,
         ));
 
         assert_ok!(Relayer::confirm_register(
@@ -354,12 +417,18 @@ fn it_provides_free_txs_confirm_done_fails_3() {
 fn it_provides_free_txs_confirm_done_fails_4() {
     new_test_ext().execute_with(|| {
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         let expected_verifying_key = BoundedVec::default();
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(5),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Public,
-            empty_program,
+            program_hash,
         ));
         let p = ValidateConfirmRegistered::<Test>::new();
         let c = RuntimeCall::Relayer(RelayerCall::confirm_register {
@@ -379,12 +448,18 @@ fn it_provides_free_txs_confirm_done_fails_4() {
 fn it_provides_free_txs_confirm_done_fails_5() {
     new_test_ext().execute_with(|| {
         let empty_program = vec![];
+        let program_hash = <Test as frame_system::Config>::Hashing::hash(&empty_program);
+        pallet_programs::Programs::<Test>::insert(
+            program_hash,
+            ProgramInfo { bytecode: empty_program, program_modification_account: 1 },
+        );
+
         let expected_verifying_key = BoundedVec::default();
         assert_ok!(Relayer::register(
             RuntimeOrigin::signed(5),
             2 as <Test as frame_system::Config>::AccountId,
             KeyVisibility::Public,
-            empty_program,
+            program_hash,
         ));
         let p = ValidateConfirmRegistered::<Test>::new();
         let c = RuntimeCall::Relayer(RelayerCall::confirm_register {