+++ runtime/src/lib.rs
@@ -1285,12 +1285,14 @@ parameter_types! {
   // 1mb max
   pub const MaxBytecodeLength: u32 = 1_000_000;
   pub const ProgramDepositPerByte: Balance = MILLICENTS;
+  pub const MaxOwnedPrograms: u32 = 250;
 }
 
 impl pallet_programs::Config for Runtime {
     type Currency = Balances;
     type MaxBytecodeLength = MaxBytecodeLength;
     type ProgramDepositPerByte = ProgramDepositPerByte;
+    type MaxOwnedPrograms = MaxOwnedPrograms;
     type RuntimeEvent = RuntimeEvent;
     type WeightInfo = weights::pallet_programs::WeightInfo<Runtime>;
 }